refactor project directory with a more organized structure
